check mountpoint

+++ mac/restore_utm.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
+
+x=$(mount |grep /Volumes/VirtualBoxImages)
+if [ -z "$x" ]; then
+  echo mount first 
+  exit 1
+fi
 rm -rf /Users/oj/Library/Containers/com.utmapp.UTM/Data/Documents/*.utm
-cp -pr /Volumes/VirtualBoxImages/UTM/ /Users/oj/Library/Containers/com.utmapp.UTM/Data/Documents/*.utm /Volumes/VirtualBoxImages/UTM/
+cp -pr /Volumes/VirtualBoxImages/UTM/*.utm /Users/oj/Library/Containers/com.utmapp.UTM/Data/Documents/
